Fetching ig works, models kinda ok, no image
