+++ backend/.env.example
@@ -1,2 +1,3 @@
 MONGO_URI=mongodb://localhost:27017/instatag
 PORT=3001
+VITE_API_BASE_URL=http://localhost:3001
