+++ backend/package.json
@@ -25,5 +25,8 @@
   "engines": {
     "node": ">=18.0.0",
     "npm": ">=8.0.0"
+  },
+"puppeteer": {
+  "default_browser_revision": ""
   }
 }
