+++ docker-compose.yml
@@ -30,6 +30,8 @@ services:
     volumes:
       - ./backend:/app
       - /app/node_modules
+    dns:
+      - 8.8.8.8
 
   mongo:
     image: mongo:7.0
