--- frontend/src/RegisterForm.css
@@ -1,46 +1,0 @@
-.register-container {
-  max-width: 400px;
-  margin: 100px auto;
-  padding: 2rem;
-  background-color: white;
-  border-radius: 12px;
-  box-shadow: 0 0 15px rgba(0,0,0,0.1);
-}
-
-h2 {
-  text-align: center;
-}
-
-form {
-  display: flex;
-  flex-direction: column;
-}
-
-input {
-  padding: 10px;
-  margin: 8px 0;
-  font-size: 16px;
-  border: 1px solid #ccc;
-  border-radius: 8px;
-  text-color:black;
-}
-
-button {
-  padding: 12px;
-  background-color: #007bff;
-  border: none;
-  color: white;
-  border-radius: 8px;
-  cursor: pointer;
-  font-size: 16px;
-}
-
-button:hover {
-  background-color: #0056b3;
-}
-
-.error {
-  color: red;
-  margin-top: 10px;
-  text-align: center;
-}
