--- frontend/src/pages/RegisterPage.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import RegisterForm from '../RegisterForm.jsx';
-
-const RegisterPage = () => {
-  return (
-    <div>
-      <h2>Rejestracja</h2>
-      <RegisterForm />
-    </div>
-  );
-};
-
-export default RegisterPage;
